+++ CHANGELOG.md
@@ -4,5 +4,6 @@
 - Options for campaign to pause, unpause and withdrawal. Also copy of campaign details to clipboard ability added
 - Added pending status to campaign while waiting for transaction complete
 - JSON Rpc urls moved to env variables
+- Added loader indicator while pause/unpause is pending
 
 ## 2.1.0